Tweaked 'convert to script' feature, cleaned a bit
This whole thing needs to be cleaned up. Its an old
mess. This was one of the first things I wrote using
GTK, or even Python for that matter. I could do much
better now. The whole cmdline/gui separation is
messy. The arg-parsing is hand-rolled (docopt is
your friend). But..it works. For me at least. :)